# docs/input/edo-backend.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ЭДО Банк — REST API (контракт + мок)",
    "version": "1.0.0",
    "description": "Спецификация бэкенд-ручек жизненного цикла обращений: регистратор, ответственный, секретарь, руководитель, аудит, уведомления, CRM.\n\n**Мок:** в `npm run dev` и `npm run preview` запросы обрабатывает Vite middleware (`src/server/edoApiMockMiddleware.ts`).\n\n**Ошибки:** передайте заголовок `X-Debug-Status` со значением `400`, `401`, `403`, `404`, `409`, `422`, `429` или `500` — вернётся соответствующий ответ с телом `application/problem+json`.\n\n**Идемпотентность:** для `POST .../actions` поддерживается заголовок `Idempotency-Key` (в моке не дедуплицируется, только документируется)."
  },
  "servers": [{ "url": "/api/v1", "description": "Текущий хост (мок Vite)" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT: `sub` — сотрудник, claim `roles` — роли (REGISTRAR, RESPONSIBLE, …)"
      }
    },
...
    "schemas": {
...
      "AppealPage": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/AppealDto" } },
          "page": { "type": "integer" },
          "size": { "type": "integer" },
          "totalElements": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Нет или невалидный Bearer",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
...
      "Validation": {
        "description": "Ошибка валидации тела или параметров",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "TooManyRequests": {
        "description": "Превышен лимит запросов",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "InternalError": {
        "description": "Внутренняя ошибка",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      }
    }
  },
  "paths": {
...
    "/me": {
      "get": {
        "tags": ["System"],
        "summary": "Профиль по токену",
        "operationId": "getMe",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Текущий пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sub": { "type": "string" },
                    "roles": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/appeals": {
      "get": {
        "tags": ["Registrar"],
        "summary": "Список обращений (регистратор / единый реестр)",
        "operationId": "listAppeals",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Поиск по номеру/ФИО" }
        ],
        "responses": {
          "200": {
            "description": "Страница обращений",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealPage" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Registrar"],
        "summary": "Создать и зарегистрировать обращение",
        "operationId": "createAppeal",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealCreate" } } }
        },
        "responses": {
          "201": {
            "description": "Создано",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealDto" } } }
          },
          "400": {
            "description": "Некорректный JSON",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/Validation" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/appeals/{appealId}": {
      "get": {
        "tags": ["Registrar"],
        "summary": "Получить обращение по id",
        "operationId": "getAppeal",
        "parameters": [{ "$ref": "#/components/parameters/AppealId" }],
        "responses": {
          "200": {
            "description": "Найдено",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealDto" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Registrar"],
        "summary": "Частичное обновление обращения",
        "operationId": "patchAppeal",
        "parameters": [{ "$ref": "#/components/parameters/AppealId" }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } }
        },
        "responses": {
          "200": {
            "description": "Обновлено",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealDto" } } }
          },
          "400": {
            "description": "Некорректный JSON",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Registrar"],
        "summary": "Удалить обращение (админ/тест)",
        "operationId": "deleteAppeal",
        "parameters": [{ "$ref": "#/components/parameters/AppealId" }],
        "responses": {
          "204": { "description": "Удалено" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/complaints/create": {
      "post": {
        "tags": ["Registrar"],
        "summary": "Внешнее создание обращения (канал API)",
        "description": "Альтернативный вход для интеграций (как в справке FAQ).",
        "operationId": "createComplaintExternal",
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealCreate" } } }
        },
        "responses": {
          "201": {
            "description": "Принято",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealDto" } } }
          },
          "400": {
            "description": "Некорректный JSON",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/Validation" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/stats/registrar/month": {
      "get": {
        "tags": ["Registrar"],
        "summary": "Статистика регистратора за месяц",
        "operationId": "getRegistrarMonthStats",
        "responses": {
          "200": {
            "description": "Агрегаты",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "withoutErrors": { "type": "integer" },
                    "withErrors": { "type": "integer" },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/responsible/appeals": {
      "get": {
        "tags": ["Responsible"],
        "summary": "Реестр обращений ответственного",
        "operationId": "listResponsibleAppeals",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" }, "style": "form", "explode": true },
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "size", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Список",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ResponsibleListResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/responsible/appeals/{appealId}": {
      "get": {
        "tags": ["Responsible"],
        "summary": "Карточка обращения (агрегат)",
        "operationId": "getResponsibleAppeal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppealId" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "header": { "$ref": "#/components/schemas/AppealDto" },
                    "actionsPreview": { "type": "array", "items": { "type": "object" } },
                    "attachments": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/responsible/appeals/{appealId}/timeline": {
      "get": {
        "tags": ["Responsible"],
        "summary": "Лента действий по обращению",
        "operationId": "getAppealTimeline",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppealId" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "type": "object" } },
                    "nextCursor": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/responsible/appeals/{appealId}/actions": {
      "post": {
        "tags": ["Responsible"],
        "summary": "Зафиксировать действие ответственного",
        "operationId": "postAppealAction",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AppealId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActionRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Действие принято",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accepted": { "type": "boolean" },
                    "appealId": { "type": "string" },
                    "type": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректный JSON",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/Validation" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/responsible/appeals/{appealId}/attachments:prepareUpload": {
      "post": {
        "tags": ["Responsible"],
        "summary": "Подготовить загрузку вложения (presigned URL)",
        "operationId": "prepareAttachmentUpload",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AppealId" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileName": { "type": "string" },
                  "mimeType": { "type": "string" },
                  "byteSize": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Данные для загрузки",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PresignedUploadResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/Validation" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/secretary/appeals": {
      "get": {
        "tags": ["Secretary"],
        "summary": "Очередь секретаря ПК (фильтр по статусу)",
        "description": "Реестр обращений на стадии ПК. **UC:** сценарии ПК в реестре UC (статус «На ПК»). **FR:** FR-02.1, FR-02.2.",
        "operationId": "listSecretaryAppeals",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "default": "На ПК" },
            "description": "Например «На ПК»"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/AppealDto" } },
                    "filter": { "type": "object" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/secretary/appeals/{appealId}": {
      "get": {
        "tags": ["Secretary"],
        "summary": "Карточка обращения для секретаря",
        "description": "**FR:** просмотр данных обращения на стадии ПК (согласовано с FR-04/FR-02 по полям карточки в UI).",
        "operationId": "getSecretaryAppeal",
        "parameters": [{ "$ref": "#/components/parameters/AppealId" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealDto" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/secretary/appeals/{appealId}/decision": {
      "post": {
        "tags": ["Secretary"],
        "summary": "Решение секретаря ПК (одобрить / отклонить)",
        "description": "Фиксация решения по обращению «На ПК». **FR:** FR-02 (переходы жизненного цикла).",
        "operationId": "postSecretaryPkDecision",
        "parameters": [{ "$ref": "#/components/parameters/AppealId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SecretaryPkDecisionRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённое обращение",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppealDto" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/Validation" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/manager/appeals": {
      "get": {
        "tags": ["Manager"],
        "summary": "Реестр обращений для руководителя",
        "description": "**UC:** UC-RU-04 (просмотр списком). **FR:** FR-14.2.",
        "operationId": "listManagerAppeals",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Поиск по номеру/ФИО/организации" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ManagerAppealsPage" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  }
}